Add support for local CIDR excludes
There are some CIDRs that should only be used locally on the host,
and never be sent to the fabric. This includes the list of 169.254
prefixes, but should exclude the metadata IP (169.254.169.254). This
patch adds support for specifying these CIDRs in a new file, which
has a .local suffix. Entries in this file will take the form:
[
    { 'src-ip': '169.254.0.0/17'},
    { 'dst-ip': '169.254.128.0/19'},
....
]

These are prefixes that require flow-mods to ensure that traffic
with this source and/or destination IP address never get sent to
the fabric.